+++ src/server/game/Chat/Commands/Level2.cpp
@@ -2215,9 +2215,9 @@ bool ChatHandler::HandlePInfoCommand(const char* args)
     }
 
     std::string timeStr = secsToTimeString(total_player_time,true,true);
-    uint32 gold = money /GOLD;
-    uint32 silv = (money % GOLD) / SILVER;
-    uint32 copp = (money % GOLD) % SILVER;
+    uint64 gold = money /GOLD;
+    uint64 silv = (money % GOLD) / SILVER;
+    uint64 copp = (money % GOLD) % SILVER;
     PSendSysMessage(LANG_PINFO_LEVEL, race_s.c_str(), Class_s.c_str(), timeStr.c_str(), level, gold, silv, copp);
 
     return true;
